Fixed Parallelization Issue with Async Invoc

+++ celery/tasks/user_mgmt_tasks.py
@@ -1,5 +1,5 @@
 import time
-
+import random
 from celery import group, states
 from shared.logger import logger
 from shared.models.user_entities import (MultipleUserIdentifiers, User,
@@ -28,15 +28,14 @@ def admin_create_user(self, *, sender: User, task_data: AddCommunityMemberReques
     logger.info(f"Adding user {task_data.email} to Auth0 at the request of {sender.email}")
     user_id = create_user(email=task_data.email, first_name=task_data.first_name,
                           last_name=task_data.last_name)
-    add_role(user_id=user_id, school=sender.school)  # allows us to identify which school a user belongs to
-    send_password_reset(email=task_data.email, first_name=task_data.first_name)
-    with Neo4JGraph() as graph:
-        graph.run("""CREATE (m: Member {first_name: $first_name, last_name: $last_name, email: $email, 
-                        location: $location, vaccinated: $vaccination, school: $school, disabled: false})""",
-                  first_name=task_data.first_name, last_name=task_data.last_name, email=task_data.email,
-                  location=task_data.location, vaccination=task_data.vaccinated, school=sender.school)
-
-    return user_id
+    if user_id:
+        add_role(user_id=user_id, school=sender.school)  # allows us to identify which school a user belongs to
+        send_password_reset(email=task_data.email, first_name=task_data.first_name)
+        with Neo4JGraph() as graph:
+            graph.run("""CREATE (m: Member {first_name: $first_name, last_name: $last_name, email: $email, 
+                            location: $location, vaccinated: $vaccination, school: $school, disabled: false})""",
+                      first_name=task_data.first_name, last_name=task_data.last_name, email=task_data.email,
+                      location=task_data.location, vaccination=task_data.vaccinated, school=sender.school)
 
 
 @celery.task(name='tasks.admin_password_reset', **GLOBAL_CELERY_OPTIONS)
@@ -65,8 +64,9 @@ def admin_delete_user(self, *, sender: User, task_data: UserIdentifier):
                   email=task_data.email, school=sender.school)
 
     user_id = get_user(email=task_data.email, fields=['user_id'])['user_id']  # get the user's user id from Auth0
-    delete_user(user_id=user_id)
-    return user_id
+
+    if user_id:
+        delete_user(user_id=user_id)
 
 
 @celery.task(name='tasks.admin_toggle_access', **GLOBAL_CELERY_OPTIONS)
@@ -77,15 +77,16 @@ def admin_toggle_access(self, *, sender: User, task_data: ToggleAccessRequest):
     :param task_data: the user identifier to toggle
     """
     logger.info(f"Setting block={task_data.block} for the user {task_data.email} at the request of {sender.email}")
-    user_id = get_user(email=task_data.email, fields=['user_id'])['user_id']
-    update_user(user_id=user_id, content={'blocked': task_data.block})
-    logger.info(f"Toggling user {task_data.email} in Neo4J")
 
     with Neo4JGraph() as graph:
         graph.run("""MATCH (m: Member {email: $email, school: $school}) SET m.disabled = $blocked""",
                   email=task_data.email, school=sender.school, blocked=task_data.block)
 
-    return user_id
+    user_id = get_user(email=task_data.email, fields=['user_id'])['user_id']
+
+    if user_id:
+        update_user(user_id=user_id, content={'blocked': task_data.block})
+        logger.info(f"Toggling user {task_data.email} in Neo4J")
 
 
 @celery.task(name='tasks.admin_bulk_import', **GLOBAL_CELERY_OPTIONS)
@@ -99,7 +100,7 @@ def admin_bulk_import(self, *, sender: User, task_data: BulkAddCommunityMemberRe
     # list of celery jobs to complete in parallel
     create_sigs = [admin_create_user.s(task_data=user, sender=sender) for user in task_data.users]
     logger.info(f"Importing {len(create_sigs)} users into MarinTrace for {sender.email}")
-    group(create_sigs).apply()
+    group(create_sigs).apply_async()
 
 
 @celery.task(name='tasks.admin_bulk_password_reset', **GLOBAL_CELERY_OPTIONS)
@@ -112,7 +113,7 @@ def admin_bulk_password_reset(self, *, sender: User, task_data: MultipleUserIden
     """
     reset_sigs = [admin_password_reset.s(task_data=identifier, sender=sender) for identifier in task_data.identifiers]
     logger.info(f"Resetting the passwords of {len(reset_sigs)} users from MarinTrace for {sender.email}")
-    group(reset_sigs).apply()
+    group(reset_sigs).apply_async()
 
 
 @celery.task(name='tasks.admin_bulk_delete_user', **GLOBAL_CELERY_OPTIONS)
@@ -125,7 +126,7 @@ def admin_bulk_delete_users(self, *, sender: User, task_data: MultipleUserIdenti
     """
     delete_sigs = [admin_delete_user.s(task_data=identifier, sender=sender) for identifier in task_data.identifiers]
     logger.info(f"Deleting {len(delete_sigs)} users from MarinTrace for {sender.email}")
-    group(delete_sigs).apply()
+    group(delete_sigs).apply_async()
 
 
 @celery.task(name='tasks.admin_bulk_toggle_access', **GLOBAL_CELERY_OPTIONS)
@@ -138,4 +139,4 @@ def admin_bulk_toggle_access(self, *, sender: User, task_data: BulkToggleAccessR
     """
     toggle_sigs = [admin_toggle_access.s(task_data=user, sender=sender) for user in task_data.users]
     logger.info(f"Changing access for {len(toggle_sigs)} users for {sender.email}")
-    group(toggle_sigs).apply()
+    group(toggle_sigs).apply_async()
